AND-8071 Make TokenItemStateConverter reusable
