+++ core/utils/src/main/java/com/tangem/utils/extensions/BigDecimalExt.kt
@@ -8,3 +8,5 @@ import java.math.BigDecimal
  * If `BigDecimal?` is `null`, returns `BigDecimal.ZERO`
  */
 fun BigDecimal?.orZero(): BigDecimal = this ?: BigDecimal.ZERO
+
+fun BigDecimal.isZero(): Boolean = this.compareTo(BigDecimal.ZERO) == 0
